updated camera UI and player view
updated the UI designs for the camera, adjusted the boundaries and the player 's camera, changed the wheel so that it turns when clicked instead of being ableto turn it left or right, and wheel also only turns left now. No longer pops up the polaroid with the image taken after the player shoots because the current method causes inconsistencies.